Add feedback controller mock
